backend/requirements.txt
asgiref==3.5.2
Django==3.2.13
django-cors-headers==3.12.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
Pillow==9.1.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
